Remove native input field on destroy
It used to just get hidden, which is bad in itself and also caused
a crash when instantiating a new one, because the NativeEditBox
instance was cleared on destruction from the m_dictReceiver on
PluginMsgHandler. When a new NativeEditBox was instantiated
the old hidden one sent an EndEditing message, because focus
was moved to the new NativeEditBox, but the receiver id could not
be found on the dictionary anymore.