Fix a deadlock, GetNativeText no longer works
We used to get a deadlock when Unity was listening to the
NativeEditBox.returnPressed event and called SetFocusNative on another
NativeEditBox on that event, which resulted in a TextEndEdit message to
be sent to Unity (which never arrived on Unity side). Not 100 % sure why
that was the case, but removing the waiting for EditBox.processRecvJsonMsg
results in SendUnityMsgToPlugin seems to solve the issue.

As we don’t wait for the result anymore, we can’t return a message to Unity.
Fortunately that was only used for the NativeEditBox.GetTextNative() method,
which isn’t necessary because we update the Unity InputField text when the
native text is changed.